requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==3.2.16
django-colorful==1.3
django-crispy-forms==1.14.0
django-filter==22.1
Faker==0.7.4
gunicorn==20.1.0
Pillow==9.4.0
python-dateutil==2.8.2
pytz==2022.1
reportlab==3.6.12
six==1.16.0
sqlparse==0.4.3
typing-extensions==4.4.0
whitenoise==6.2.0
